+++ data/normalized/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Latest Microsoft Security News Changes
 
-Generated: 2026-08-16T06:33:20.215Z
+Generated: 2026-08-17T06:47:14.218Z
 
 - 2026-08-01 • Microsoft Defender for Cloud • Preview • Date Category Update August 6, 2026 Public preview On-demand malware scanning now
 - 2026-08-01 • Microsoft Defender for Endpoint • Preview • Preview/GA
@@ -11,6 +11,7 @@ Generated: 2026-08-16T06:33:20.215Z
 - 2026-08-01 • Microsoft Sentinel • Preview • UEBA anomalies on behaviors (Preview)
 - 2026-08-01 • Microsoft Sentinel • Ikke oppgitt • UEBA anomaly detection now supports Check Point, Fortinet FortiGate, and Zscaler firewall, VPN, and web proxy events from the
 - 2026-08-01 • Microsoft Sentinel • Ikke oppgitt • UEBA anomaly detection also supports identity-linked AWS GuardDuty findings from the
+- 2026-08-01 • Security Exposure Management • Ikke oppgitt • Codename MDASH - Cancel scan
 - 2026-07-01 • Microsoft Defender for Cloud • Deprecation • Deprecation completed: Legacy grouped recommendations removed from Azure portal
 - 2026-07-01 • Microsoft Defender for Cloud • GA • GA Database-level recommendations for SQL Vulnerability Assessment (GA) July 5, 2026
 - 2026-07-01 • Microsoft Defender for Cloud • Deprecation • Deprecation
@@ -21,4 +22,3 @@ Generated: 2026-08-16T06:33:20.215Z
 - 2026-07-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 1.1.78290102 released: See enhancements and features for this release
 - 2026-07-01 • Microsoft Defender for Endpoint • Preview • Preview
 - 2026-07-01 • Microsoft Defender for Endpoint • Ikke oppgitt • GA Release version 20.126052.16.0 released: See enhancements and features for this release
-- 2026-07-01 • Microsoft Defender for Identity • Ikke oppgitt • Defender for Identity sensor updates